Make cli easier to use as a library, fix a couple warnings
